Abstract chassis class
