+++ include/lemlib/api.hpp
@@ -13,6 +13,6 @@
 #include "lemlib/util.hpp"
 #include "lemlib/pid.hpp"
 #include "lemlib/pose.hpp"
-#include "lemlib/chassis/chassis.hpp"
+#include "lemlib/chassis/differential.hpp"
 
 #include "lemlib/logger/logger.hpp"
